requirements.txt
requests
pycryptodome
websocket-client
numpy>=1.26
mss>=9.0.1
Pillow>=10.2.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
pytest>=8.1.1
